requirements.txt
black==24.4.2
celery==5.4.0
Django==5.0.6
django-celery-beat==2.6.0
django-cleanup==8.1.0
django-simple-history==3.7.0
pillow==10.3.0
redis==5.0.6